Fix falling ent displacing hot item crash
To repro: place a sand node at least 2 nodes high above
glowing lode prills and let the sand fall onto the prills and
try to displace them.

Introduced in 01252653-5da0918.

For some reason it seems like a non-player object is being
passed as the digger (I thought it would be nil for non-
players?) and this object apparently imitates a player but
is missing some critical thing such as metadata.

To work around the problem, just add guards to the
phealth functions, since only players are intended to have
"health" anyway.