Remove emergency digging
This was added to allow a player to get out of a
rare softlock (trapped inside a cobble room without
any tools).  Since then we have not seen much use
of it for this intended purpose.

However, it has caused MUCH confusion among new
players, wondering why it takes so long to mine
dirt or tree trunks, instead of trying to figure out how
to make the proper tool.

Emergency digging is causing more problems than
it's solving and thus it makes more sense to remove
it than to keep it.  If we need some mechanic to fill
its niche (i.e. a way out of softlock) then we will need
to consider something that's harder to discover
accidentally to avoid confusion.