Move node itemstacks from meta inv to field
Node inventories are sent to clients, while metadata
fields are automatically set private by the server
and not sent to clients.  Not using node meta
inventories may help reduce the amount of data
sent to clients, especially in storage-heavy areas
where we were duplicating the work of sending the
inventory data not used on the client, as well as the
visible item entities.

Automatically convert old-format data to new upon
reading.

Also, needed to provide some APIs for sane access
to data in serialized format, and another way to
signal that a node can accept a stack other than
the size of inventory[solo]