Workaround pseudo-rubberbanding issue
Old behavior used to round exactly-off-by-half
values up toward +inf, but since
https://github.com/minetest/minetest/pull/10803
they are rounded toward +/-inf, which means that
if a player's y coord was exactly -1.5, it used to
round up to -1.0 but now rounds down to -2.0.

This affects using vector.round to get the pos of
the node the player's feet are standing INSIDE,
because if the player's y coord is exactly -0.5, as
would happen while they're standing perfectly at
rest on top of a node, it will instead give us the
node they're standing ON.

This causes spurious player "pushout" events,
because the game thinks players are standing
inside the node they're actually standing on top
of.

To work around this, just add a small positive
fudge factor to the player's y coord to bias it
toward +inf.