Display when cheats are enabled on-screen
Without this, some players are likely to enable
certain privs and abilities without knowing that
they're cheats and will break the designed
gameplay.  Similarly, server admins may be likely
to use the same default_privs for a nodecore server
as they might for an MTG or even creative server,
causing accidental cheating as well.  Adding this
messages makes it hard to be unaware that your
game is at least misconfigured.