Dynamic custom crosshair
- Crosshair hidden when !interact
- Fade out crosshair when not pointing
- Makes crosshair HUD-scalable (not sure if
  this is definitely good or bad yet)
- Forces crosshair appearance on mobile to
  match desktop

Since the crosshair position is already used in
gameplay i.e. for item dropping on mobile, it's
important that mobile have consistent access
to it compared to desktop.

Also, mobile's lack of crosshair was one of the
main blockers to having look-tips instead of
touch-tips, so forcing a crosshair opens that
up again.  Forcing punch interactions to see
anything causes issues with ember/pumwater
where punching hurts.

A crosshair would also be necessary for
star navigation assistance tools.