Fix loose self-repacking when sitting on top of loose nodes
The original idea was that stacks of loose nodes should repack
from the bottom up, and self-repacking should never "race" against
reposing.

If the node is sitting on a solid node, it can always self-repack.
If it's sitting on a loose node, then only repack if the node
below is not itself in the process of self-repacking, and if the
node itself isn't awaiting reposing.