Crosshair shape by protection instead of type
The use of object vs node crosshair is less important now
due to our ability to rely on the looktips to differentiate
which thing we're pointing at (and which face), and
makes less sense in terms of NodeCore's everything-is-a-
node nature.

Protection, however, is more useful to show, where in
use, so players can avoid interacting with things and
triggering protection alerts.  Use the old everted
crosshair design to now mean "protected" instead of
"pointing at entity."

The old use of the crosshair shape should have been
uncommon enough that players should be able to
adapt to the change alright.