Try to mitigate spammy looktips
Right now looktips are rather obtrusive to
the overall game experience, and should
show up on some kind of delay to mitigate
this.

I tried having a delay for looking at a node
before a looktip showed up, but it felt sort
of annoying when switching from node to
node, especially since the faster response
on the face-dependent location of the tip
is helpful for building.  The code for it also
turned out to be messy, since node and
object handling needed to be different, since
it was unrealistic to actually hold the
crosshair on a moving thrown object.

What I've got for now is just disabling the
looktip during, and for a slight delay right
after, any player movement, based on
movement input controls.  This prevents
the player from having a bunch of looktips
juddering across the screen while walking,
but allows them whenever standing still.

This MAY be good enough, but it may take
some play-testing to confirm.