Rebrand "hints" as "challenges"
There seems to be a common misconception among
new players that the "hint" system is supposed to
help provide guidance when they're stuck, when what
they're really for is hinting that something exists and
challenging the player to find out how to access it.

Change the name of "hints" to "challenges" for all
text display purposes to try to clarify.