Add some warnings to the readme
It seems like a number of players are coming in with
misconceptions about the type of game this is or
what the experience is supposed to be like.  Add
some explicit warnings to the CDB long description to
make it clearer.

If these help, it may make sense to add similar
warnings to the player's guide.