Remember player's last inventory tab
Note that some tabs may have time-dependent contents (e.g. the
hints/challenges tab, or add-on mod tabs) and need to be notified
of this in the event that they are the current tab.  It will
be the responsibility of each tab to watch for appropriate events
and perform background updates as necessary.  The hint tab does
this on an event-driven basis from player discovery.