Fix stack ignition conservation of mass
This was a workaround for a bug that it never actually really
addressed that caused stack fires to rage out of control.  It
turned out the bug was a special case item duplication, and it
was actually fixed long ago.

Allow stacks of things to be burned for all their embers/ash
again, and also for more devastating shelf-fire disasters.