Storeboxes: don't fall in unless you can fall thru
There isn't always a surface below to catch items
that fall into a storebox, so we can't guarantee an
item that falls into a storebox won't just fall out
the bottom.  Only allow falling in the top if falling
out the bottom is allowed too.

This fixes glass cases not catching items if the
case is floating in air with nothing solid below it.

There are still some cases it doesn't handle
correctly, like a bare stack node underneath a
form, wherein the item should be allowed to fall
out of the form and combine into the stack.