Fix bandolier in stand, wave, mine anims
MT seems to have a problem with bones that are not animated
and have exactly 0 rotation, i.e. interpreting this as 180 degrees
(supposedly this is a workaround for some other quirk).  Give
the chest and bandolier bones tiny non-zero rotation adjustments
to trick MT into putting them as close as possible to zero without
triggering this bug.