Fix quench check on heating recipes
If there's any coolant near a node, that blocks
all application of heat and prevents it from being
cooked, and also prevents any passive slow-cooling
recipes from operating on it (i.e. racing against
the quenched recipes).

Use coolant neargroups for quench check too,
for consistency.  Add memoization to the recipe
data on the check, so to avoid doing the quench
scan for each recipe separately.