Make pumwater much more common
There is a chance to find some at any depth.  At
around -256, it should start to get practical to
actually hunt for it.  By -600 or so, it is common
enough to actually be an obstacle, and deep
caverns start to get pum-filled.

Pumwater is added by scattering "blob" ore
throughout stone at various depths.  Stone
immediately above pumwater sources is carved
away to air, to encourage the pumwater ambiance
to play, to make it a bit easier to find it by audio
cues.