Make machines always do selective raking
Machines can only dig one way, and cannot hold the "sneak" key,
nor is there a particularly sensible way to differentiate between
different kind of actions or setup the environment to signal
the digging "mode".

Toward this end, it seems to make the most sense to have the thing
that the machine is digging always significant, and it should be
possible to emulate a "dig everything" machine (knowing that only
certain kinds of things will be within range within a machine)
using multiple "dig one thing" rakes, whereas the reverse is not
so obvious.

For this reason, until a multi-modal system is figured out, change
the default mode to be selective.