Cache player interact privs for current tick
Apparently accessing player privileges is
weirdly expensive, regardless of whether
get_player_privs or check_player_privs is used,
and because interact privs are checked so
frequently, it adds up to quite a lot (as much as
a third of the playerstep time).  Since things
controlled by interact privs are probably safe
to delay for a tick (it won't be toggled very
often), just cache it for the current tick to
mitigate this.

In the future it might make sense to (1) trap
the privilege change calls to invalidate the
cache, and (2) consider  doing this for some
other privilege checks.