Fix inefficient dynamic light checks
Instead of checking to make sure one of the objects
found at a location is the glowing entity (which is
very slow and has to search every object for each
light) just make sure that the entity still exists
(get_pos() does not return nil) and that it's still
within the same node position as before.

This fixes a huge lagspike that was caught on NCC
due to a bunch of dropped glowing lenses.