Fix off-by-zero error in dnt timer setting
If the previously set timer is exactly the correct
value for the existing DNT, then don't update/reset
the node timer.

Not sure why this was the way it was before (it
seems like it might have been intentional) but it
caused measurable performance impact, didn't
seem right, and fixing it didn't seem to introduce
new problems in preliminary testing.