Allow optic pulsers to trigger ablation sooner
Problems seem to arise when using a door
for an optic timer to produce exact 2 second
cycles.  Timing jank can cause the cycle to
trigger faster than downstream doors are
actually ready to trigger, causing some doors
to fail to fire, causing machine desync errors.
Try allowing ablation to refire sooner (but not
immediately) if the power is removed from the
lens pushing the door.