Try a different method to register brick doors
A heisenbug has been observed causing some
of the concrete brick doors not to be registered
on startups randomly/rarely.  No cause has
been found yet, though it could be an issue
with MT's mod loading order; this is the only
hypothesis that seems consistent with the
existing code.

Try using register_on_register_item instead of
just looping through currently-existing defs, to
see if this fixes it.