Snapshot items before iterating in retroactive on_register
One potential source of the "missed item registration" bug could
be the on_register callback running multiple times for an item if
it was registered during a retroactive on_register scan.  The
behavior of pairs() if the table being iterated is modified is
apparently undefined.  In fact, it may be possible for a table to
get rehashed, such that some items might be skipped.  Taking
a snapshot first and iterating that instead should be more
stable overall.